# Conflicts:
#	OpenStreetMapViewer/src/main/java/org/osmdroid/samplefragments/tileproviders/GeopackageSample.java
#	OpenStreetMapViewer/src/main/java/org/osmdroid/samplefragments/tileproviders/MapsforgeTileProviderSample.java
#	osmdroid-android/src/main/java/org/osmdroid/api/IMapController.java
#	osmdroid-android/src/main/java/org/osmdroid/views/MapController.java

+++ osmdroid-android/src/main/java/org/osmdroid/views/drawing/OsmPath.java
@@ -21,7 +21,7 @@ public class OsmPath extends Path {
 
 	private final static GeoPoint sReferenceGeoPoint = new GeoPoint(0, 0);
 	protected final Point mReferencePoint = new Point();
-	private int mLastZoomLevel = -1;
+	private double mLastZoomLevel = -1;
 	
     public OsmPath() {
 		super();
